+++ pytorch_lightning/metrics/functional/stat_scores.py
@@ -260,9 +260,9 @@ def stat_scores(
         >>> stat_scores(preds, target, reduce='macro', num_classes=3)
         tensor([[0, 1, 2, 1, 1],
                 [1, 1, 1, 1, 2],
-                [1, 0, 3, 0, 1]], dtype=torch.int32)
+                [1, 0, 3, 0, 1]])
         >>> stat_scores(preds, target, reduce='micro')
-        tensor([2, 2, 6, 2, 4], dtype=torch.int32)
+        tensor([2, 2, 6, 2, 4])
 
     """
